Oleksandr_Khomenko/summarization-agent/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
openai>=1.10.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
scikeras>=0.12.0
scikit-learn>=1.3.0
transformers>=4.30.0
torch>=2.0.0
summarizers>=0.0.7
pillow>=10.0.0
matplotlib>=3.7.0
sumy>=0.11.0
nltk>=3.8.1
